Plane: use mission loiter radius when checking target waypoint
The loiter algorithm checks to see if the next waypoint is within the loiter
radius, and if so, breaks out of the loiter (since the bearing will never line
up with the target). Before this patch, this check always used WP_LOITER_RAD,
even if the mission item had a different loiter radius.

This caused loiters to be skipped if their radius was smaller than WP_LOITER_RAD
and there was a waypoint less than WP_LOITER_RAD from the center. Alternatively,
if the mission radius was larger than WP_LOITER_RAD and the next waypoint was
placed between this radius and WP_LOITER_RAD, the vehicle would complete several
rotations before the heading tolerance increased enough to trigger the exit.