AP_TECS: reset throttle integrator when not using airspeed sensor
The throttle integrator is only with an airspeed sensor, therefore I only use
it during flare. It tends to wind up after touchdown, but normally gets reset
when switching away from an auto-throttle mode. If we stay in auto though, it
will not reset and mess up the next landing.