RCOutput_Bebop: move ESC telemetry off of background thread
Read the ESC telemetry directly in the main scheduler loop. The background loop
appears to primarily be used to make sure the ESC watchdog doesn't time out if
the motors stop being updated. I don't see a reason the ESC telemetry can't be
in the main loop instead, and this avoids some locks. In fact, I'm not sure the
background thread is necessary at all.